requirements.txt
fpdf2
fastapi
docker
markdown2